app/requirements.txt
azure-identity
psycopg2-binary
